+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Added a `NIP98` module to verify NIP-98 Requests.
+
 ## [0.21.1] - 2024-05-18
 
 ## Changed